+++ backend/ee/onyx/background/task_name_builders.py
@@ -6,7 +6,9 @@ from onyx.configs.constants import OnyxCeleryTask
 QUERY_HISTORY_TASK_NAME_PREFIX = OnyxCeleryTask.EXPORT_QUERY_HISTORY_TASK
 
 
-def name_chat_ttl_task(retention_limit_days: int, tenant_id: str | None = None) -> str:
+def name_chat_ttl_task(
+    retention_limit_days: float, tenant_id: str | None = None
+) -> str:
     return f"chat_ttl_{retention_limit_days}_days"
