+++ backend/onyx/server/settings/models.py
@@ -43,7 +43,8 @@ class Notification(BaseModel):
 class Settings(BaseModel):
     """General settings"""
 
-    maximum_chat_retention_days: int | None = None
+    # is float to allow for fractional days for easier automated testing
+    maximum_chat_retention_days: float | None = None
     gpu_enabled: bool | None = None
     application_status: ApplicationStatus = ApplicationStatus.ACTIVE
     anonymous_user_enabled: bool | None = None
